requirements.txt
fbchat-muqit 
google-generativeai
python-dotenv
cryptography
